refactor(oracle): migrate ble with whole-stack raw-bytes encoder
